+++ source/libs/function/src/detail/tminmax.c
@@ -72,173 +72,6 @@
 
 #define GET_INVOKE_INTRINSIC_THRESHOLD(_bits, _bytes) ((_bits) / ((_bytes) << 3u))
 
-#ifdef __AVX2__
-static void calculateRounds(int32_t numOfRows, int32_t bytes, int32_t* remainder, int32_t* rounds, int32_t* width) {
-  const int32_t bitWidth = 256;
-
-  *width = (bitWidth >> 3u) / bytes;
-  *remainder = numOfRows % (*width);
-  *rounds = numOfRows / (*width);
-}
-
-#define EXTRACT_MAX_VAL(_first, _sec, _width, _remain, _v) \
-  __COMPARE_EXTRACT_MAX(0, (_width), (_v), (_first))       \
-  __COMPARE_EXTRACT_MAX(0, (_remain), (_v), (_sec))
-
-#define EXTRACT_MIN_VAL(_first, _sec, _width, _remain, _v) \
-  __COMPARE_EXTRACT_MIN(0, (_width), (_v), (_first))       \
-  __COMPARE_EXTRACT_MIN(0, (_remain), (_v), (_sec))
-
-#define CMP_TYPE_MIN_MAX(type, cmp)                      \
-  const type* p = pData;                                 \
-  __m256i     initVal = _mm256_lddqu_si256((__m256i*)p); \
-  p += width;                                            \
-  for (int32_t i = 1; i < (rounds); ++i) {               \
-    __m256i next = _mm256_lddqu_si256((__m256i*)p);      \
-    initVal = CMP_FUNC_##cmp##_##type(initVal, next);    \
-    p += width;                                          \
-  }                                                      \
-  const type* q = (const type*)&initVal;                 \
-  type*       v = (type*)res;                            \
-  EXTRACT_##cmp##_VAL(q, p, width, remain, *v)
-
-static void i8VectorCmpAVX2(const void* pData, int32_t numOfRows, bool isMinFunc, bool signVal, int64_t* res) {
-  const int8_t* p = pData;
-
-  int32_t width, remain, rounds;
-  calculateRounds(numOfRows, sizeof(int8_t), &remain, &rounds, &width);
-
-#define CMP_FUNC_MIN_int8_t  _mm256_min_epi8
-#define CMP_FUNC_MAX_int8_t  _mm256_max_epi8
-#define CMP_FUNC_MIN_uint8_t _mm256_min_epu8
-#define CMP_FUNC_MAX_uint8_t _mm256_max_epu8
-
-  if (!isMinFunc) {  // max function
-    if (signVal) {
-      CMP_TYPE_MIN_MAX(int8_t, MAX);
-    } else {
-      CMP_TYPE_MIN_MAX(uint8_t, MAX);
-    }
-  } else {  // min function
-    if (signVal) {
-      CMP_TYPE_MIN_MAX(int8_t, MIN);
-    } else {
-      CMP_TYPE_MIN_MAX(uint8_t, MIN);
-    }
-  }
-}
-
-static void i16VectorCmpAVX2(const void* pData, int32_t numOfRows, bool isMinFunc, bool signVal, int64_t* res) {
-  int32_t width, remain, rounds;
-  calculateRounds(numOfRows, sizeof(int16_t), &remain, &rounds, &width);
-
-#define CMP_FUNC_MIN_int16_t  _mm256_min_epi16
-#define CMP_FUNC_MAX_int16_t  _mm256_max_epi16
-#define CMP_FUNC_MIN_uint16_t _mm256_min_epu16
-#define CMP_FUNC_MAX_uint16_t _mm256_max_epu16
-  if (!isMinFunc) {  // max function
-    if (signVal) {
-      CMP_TYPE_MIN_MAX(int16_t, MAX);
-    } else {
-      CMP_TYPE_MIN_MAX(uint16_t, MAX);
-    }
-  } else {  // min function
-    if (signVal) {
-      CMP_TYPE_MIN_MAX(int16_t, MIN);
-    } else {
-      CMP_TYPE_MIN_MAX(uint16_t, MIN);
-    }
-  }
-}
-
-static void i32VectorCmpAVX2(const void* pData, int32_t numOfRows, bool isMinFunc, bool signVal, int64_t* res) {
-  int32_t width, remain, rounds;
-  calculateRounds(numOfRows, sizeof(int32_t), &remain, &rounds, &width);
-
-#define CMP_FUNC_MIN_int32_t  _mm256_min_epi32
-#define CMP_FUNC_MAX_int32_t  _mm256_max_epi32
-#define CMP_FUNC_MIN_uint32_t _mm256_min_epu32
-#define CMP_FUNC_MAX_uint32_t _mm256_max_epu32
-  if (!isMinFunc) {  // max function
-    if (signVal) {
-      CMP_TYPE_MIN_MAX(int32_t, MAX);
-    } else {
-      CMP_TYPE_MIN_MAX(uint32_t, MAX);
-    }
-  } else {  // min function
-    if (signVal) {
-      CMP_TYPE_MIN_MAX(int32_t, MIN);
-    } else {
-      CMP_TYPE_MIN_MAX(uint32_t, MIN);
-    }
-  }
-}
-
-static void floatVectorCmpAVX2(const float* pData, int32_t numOfRows, bool isMinFunc, float* res) {
-  const float* p = pData;
-
-  int32_t width, remain, rounds;
-  calculateRounds(numOfRows, sizeof(float), &remain, &rounds, &width);
-
-  __m256 next;
-  __m256 initVal = _mm256_loadu_ps(p);
-  p += width;
-
-  if (!isMinFunc) {  // max function
-    for (int32_t i = 1; i < rounds; ++i) {
-      next = _mm256_loadu_ps(p);
-      initVal = _mm256_max_ps(initVal, next);
-      p += width;
-    }
-
-    const float* q = (const float*)&initVal;
-    EXTRACT_MAX_VAL(q, p, width, remain, *res)
-  } else {  // min function
-    for (int32_t i = 1; i < rounds; ++i) {
-      next = _mm256_loadu_ps(p);
-      initVal = _mm256_min_ps(initVal, next);
-      p += width;
-    }
-
-    const float* q = (const float*)&initVal;
-    EXTRACT_MIN_VAL(q, p, width, remain, *res)
-  }
-}
-
-static void doubleVectorCmpAVX2(const double* pData, int32_t numOfRows, bool isMinFunc, double* res) {
-  const double* p = pData;
-
-  int32_t width, remain, rounds;
-  calculateRounds(numOfRows, sizeof(double), &remain, &rounds, &width);
-
-  __m256d next;
-  __m256d initVal = _mm256_loadu_pd(p);
-  p += width;
-
-  if (!isMinFunc) {  // max function
-    for (int32_t i = 1; i < rounds; ++i) {
-      next = _mm256_loadu_pd(p);
-      initVal = _mm256_max_pd(initVal, next);
-      p += width;
-    }
-
-    // let sum up the final results
-    const double* q = (const double*)&initVal;
-    EXTRACT_MAX_VAL(q, p, width, remain, *res)
-  } else {  // min function
-    for (int32_t i = 1; i < rounds; ++i) {
-      next = _mm256_loadu_pd(p);
-      initVal = _mm256_min_pd(initVal, next);
-      p += width;
-    }
-
-    // let sum up the final results
-    const double* q = (const double*)&initVal;
-    EXTRACT_MIN_VAL(q, p, width, remain, *res)
-  }
-}
-#endif
-
 static int32_t findFirstValPosition(const SColumnInfoData* pCol, int32_t start, int32_t numOfRows, bool isStr) {
   int32_t i = start;
 
@@ -255,31 +88,31 @@ static void handleInt8Col(const void* data, int32_t start, int32_t numOfRows, SM
     pBuf->v = ((const int8_t*)data)[start];
   }
 
-#ifdef __AVX2__
-  if (tsAVX2Supported && tsSIMDEnable && numOfRows * sizeof(int8_t) >= sizeof(__m256i)) {
-    i8VectorCmpAVX2(data + start * sizeof(int8_t), numOfRows, isMinFunc, signVal, &pBuf->v);
-  } else {
-#else
-  if (true) {
-#endif
-    if (signVal) {
-      const int8_t* p = (const int8_t*)data;
-      int8_t*       v = (int8_t*)&pBuf->v;
+  if (tsAVX2Supported && tsSIMDEnable && numOfRows * sizeof(int8_t) >= M256_BYTES) {
+    int32_t code = i8VectorCmpAVX2(((char*)data) + start * sizeof(int8_t), numOfRows, isMinFunc, signVal, &pBuf->v);
+    if (code == TSDB_CODE_SUCCESS) {
+      pBuf->assign = true;
+      return;
+    }
+  }
 
-      if (isMinFunc) {
-        __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
-      } else {
-        __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
-      }
+  if (signVal) {
+    const int8_t* p = (const int8_t*)data;
+    int8_t*       v = (int8_t*)&pBuf->v;
+
+    if (isMinFunc) {
+      __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
     } else {
-      const uint8_t* p = (const uint8_t*)data;
-      uint8_t*       v = (uint8_t*)&pBuf->v;
+      __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
+    }
+  } else {
+    const uint8_t* p = (const uint8_t*)data;
+    uint8_t*       v = (uint8_t*)&pBuf->v;
 
-      if (isMinFunc) {
-        __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
-      } else {
-        __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
-      }
+    if (isMinFunc) {
+      __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
+    } else {
+      __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
     }
   }
 
@@ -292,31 +125,31 @@ static void handleInt16Col(const void* data, int32_t start, int32_t numOfRows, S
     pBuf->v = ((const int16_t*)data)[start];
   }
 
-#ifdef __AVX2__
-  if (tsAVX2Supported && tsSIMDEnable && numOfRows * sizeof(int16_t) >= sizeof(__m256i)) {
-    i16VectorCmpAVX2(data + start * sizeof(int16_t), numOfRows, isMinFunc, signVal, &pBuf->v);
-  } else {
-#else
-  if (true) {
-#endif
-    if (signVal) {
-      const int16_t* p = (const int16_t*)data;
-      int16_t*       v = (int16_t*)&pBuf->v;
+  if (tsAVX2Supported && tsSIMDEnable && numOfRows * sizeof(int16_t) >= M256_BYTES) {
+    int32_t code = i16VectorCmpAVX2(((char*)data) + start * sizeof(int16_t), numOfRows, isMinFunc, signVal, &pBuf->v);
+    if (code == TSDB_CODE_SUCCESS) {
+      pBuf->assign = true;
+      return;
+    }
+  }
 
-      if (isMinFunc) {
-        __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
-      } else {
-        __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
-      }
+  if (signVal) {
+    const int16_t* p = (const int16_t*)data;
+    int16_t*       v = (int16_t*)&pBuf->v;
+
+    if (isMinFunc) {
+      __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
     } else {
-      const uint16_t* p = (const uint16_t*)data;
-      uint16_t*       v = (uint16_t*)&pBuf->v;
+      __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
+    }
+  } else {
+    const uint16_t* p = (const uint16_t*)data;
+    uint16_t*       v = (uint16_t*)&pBuf->v;
 
-      if (isMinFunc) {
-        __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
-      } else {
-        __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
-      }
+    if (isMinFunc) {
+      __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
+    } else {
+      __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
     }
   }
 
@@ -329,31 +162,31 @@ static void handleInt32Col(const void* data, int32_t start, int32_t numOfRows, S
     pBuf->v = ((const int32_t*)data)[start];
   }
 
-#ifdef __AVX2__
-  if (tsAVX2Supported && tsSIMDEnable && numOfRows * sizeof(int32_t) >= sizeof(__m256i)) {
-    i32VectorCmpAVX2(data + start * sizeof(int32_t), numOfRows, isMinFunc, signVal, &pBuf->v);
-  } else {
-#else
-  if (true) {
-#endif
-    if (signVal) {
-      const int32_t* p = (const int32_t*)data;
-      int32_t*       v = (int32_t*)&pBuf->v;
+  if (tsAVX2Supported && tsSIMDEnable && numOfRows * sizeof(int32_t) >= M256_BYTES) {
+    int32_t code = i32VectorCmpAVX2(((char*)data) + start * sizeof(int32_t), numOfRows, isMinFunc, signVal, &pBuf->v);
+    if (code == TSDB_CODE_SUCCESS) {
+      pBuf->assign = true;
+      return;
+    }
+  }
 
-      if (isMinFunc) {
-        __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
-      } else {
-        __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
-      }
+  if (signVal) {
+    const int32_t* p = (const int32_t*)data;
+    int32_t*       v = (int32_t*)&pBuf->v;
+
+    if (isMinFunc) {
+      __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
     } else {
-      const uint32_t* p = (const uint32_t*)data;
-      uint32_t*       v = (uint32_t*)&pBuf->v;
+      __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
+    }
+  } else {
+    const uint32_t* p = (const uint32_t*)data;
+    uint32_t*       v = (uint32_t*)&pBuf->v;
 
-      if (isMinFunc) {
-        __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
-      } else {
-        __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
-      }
+    if (isMinFunc) {
+      __COMPARE_EXTRACT_MIN(start, start + numOfRows, *v, p);
+    } else {
+      __COMPARE_EXTRACT_MAX(start, start + numOfRows, *v, p);
     }
   }
 
@@ -397,18 +230,18 @@ static void handleFloatCol(SColumnInfoData* pCol, int32_t start, int32_t numOfRo
     *val = pData[start];
   }
 
-#ifdef __AVX2__
-  if (tsAVXSupported && tsSIMDEnable && numOfRows * sizeof(float) >= sizeof(__m256i)) {
-    floatVectorCmpAVX2(pData + start, numOfRows, isMinFunc, val);
-  } else {
-#else
-  if (true) {
-#endif
-    if (isMinFunc) {  // min
-      __COMPARE_EXTRACT_MIN(start, start + numOfRows, *val, pData);
-    } else {  // max
-      __COMPARE_EXTRACT_MAX(start, start + numOfRows, *val, pData);
+  if (tsAVX2Supported && tsSIMDEnable && numOfRows * sizeof(float) >= M256_BYTES) {
+    int32_t code = floatVectorCmpAVX2(pData + start, numOfRows, isMinFunc, val);
+    if (code == TSDB_CODE_SUCCESS) {
+      pBuf->assign = true;
+      return;
     }
+  }
+
+  if (isMinFunc) {  // min
+    __COMPARE_EXTRACT_MIN(start, start + numOfRows, *val, pData);
+  } else {  // max
+    __COMPARE_EXTRACT_MAX(start, start + numOfRows, *val, pData);
   }
 
   pBuf->assign = true;
@@ -422,18 +255,18 @@ static void handleDoubleCol(SColumnInfoData* pCol, int32_t start, int32_t numOfR
     *val = pData[start];
   }
 
-#ifdef __AVX2__
-  if (tsAVXSupported && tsSIMDEnable && numOfRows * sizeof(double) >= sizeof(__m256i)) {
-    doubleVectorCmpAVX2(pData + start, numOfRows, isMinFunc, val);
-  } else {
-#else
-  if (true) {
-#endif
-    if (isMinFunc) {  // min
-      __COMPARE_EXTRACT_MIN(start, start + numOfRows, *val, pData);
-    } else {  // max
-      __COMPARE_EXTRACT_MAX(start, start + numOfRows, *val, pData);
+  if (tsAVX2Supported && tsSIMDEnable && numOfRows * sizeof(double) >= M256_BYTES) {
+    int32_t code = doubleVectorCmpAVX2(pData + start, numOfRows, isMinFunc, val);
+    if (code == TSDB_CODE_SUCCESS) {
+      pBuf->assign = true;
+      return;
     }
+  }
+
+  if (isMinFunc) {  // min
+    __COMPARE_EXTRACT_MIN(start, start + numOfRows, *val, pData);
+  } else {  // max
+    __COMPARE_EXTRACT_MAX(start, start + numOfRows, *val, pData);
   }
 
   pBuf->assign = true;
